backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.9
django-admin-tools==0.9.2
django-nested-admin==3.3.3
django-nested-inline==0.4.4
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
idna==3.3
Markdown==3.3.4
martor==1.6.9
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-monkey-business==1.0.0
pytz==2021.3
requests==2.27.1
routers==0.10.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
